refactor: change packages from plural to singular
