+++ src/main/java/com/github/dougmab/openvinylboxapi/controller/CategoryController.java
@@ -1,8 +1,7 @@
-package com.github.dougmab.openvinylboxapi.controllers;
+package com.github.dougmab.openvinylboxapi.controller;
 
 import com.github.dougmab.openvinylboxapi.dto.CategoryDTO;
-import com.github.dougmab.openvinylboxapi.entities.Category;
-import com.github.dougmab.openvinylboxapi.services.CategoryService;
+import com.github.dougmab.openvinylboxapi.service.CategoryService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -10,7 +9,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @RestController
