+++ src/main/java/com/github/dougmab/openvinylboxapi/controller/exception/ControllerExceptionHandler.java
@@ -1,4 +1,4 @@
-package com.github.dougmab.openvinylboxapi.controllers.exceptions;
+package com.github.dougmab.openvinylboxapi.controller.exception;
 
 import jakarta.persistence.EntityNotFoundException;
 import jakarta.servlet.http.HttpServletRequest;
@@ -6,8 +6,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-
-import java.time.Instant;
 
 @ControllerAdvice
 public class ControllerExceptionHandler {
